+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.40] - 2025-11-26
+### Changed:
+- Moved all `pandas` related functions into a specific `.helpers.pandas` submodule.
+### Added:
+- `.helpers.pandas.search` function to search within a `pandas` dataframe.
+
 ## [0.0.39] - 2025-10-30
 ### Added:
 - Added `Modeling.fit_negbin` to model response that follows negative binomial distribution.